finished build

+++ src/Customer.cpp
@@ -26,7 +26,7 @@ std::vector<int> SweatyCustomer::order(const std::vector<Workout> &workout_optio
     // filteredWorkouts only lives in this scope - therefore on stack;
     // option 2 initial in heap and delete before function close.
     std::vector<int> filteredWorkouts;
-    for(int i = 0; i < workout_options.size(); ++i){
+    for(size_t i = 0; i < workout_options.size(); ++i){
         if(workout_options[i].getType() == CARDIO){
             filteredWorkouts.push_back(workout_options[i].getId());
         }
@@ -51,7 +51,7 @@ std::vector<int> CheapCustomer::order(const std::vector<Workout> &workout_option
     // we dereference cheapest to hold an alias to the first workout element in the list.
     const Workout *cheapest = &(workout_options[0]);
     std::vector<int> filteredWorkouts;
-    for(int i = 1; i < workout_options.size(); ++i){
+    for(size_t i = 1; i < workout_options.size(); ++i){
         if(workout_options[i].getPrice() < cheapest -> getPrice()){
             //changing the cheapest alias to another object.
             cheapest = &(workout_options[i]);
@@ -76,7 +76,7 @@ std::vector<int> HeavyMuscleCustomer::order(const std::vector<Workout> &workout_
     std::vector<int> sortedWorkouts;
     //array of pointers to the filtered workouts.
     std::vector<const Workout*> filteredWorkouts;
-    for(int i = 0; i <workout_options.size(); i++){
+    for(size_t i = 0; i <workout_options.size(); i++){
         if(workout_options[i].getType() == ANAEROBIC){
             filteredWorkouts.push_back(&(workout_options[i]));
         }
@@ -110,7 +110,7 @@ std::vector<int> FullBodyCustomer::order(const std::vector<Workout> &workout_opt
     const Workout *cheapestAnaerobic = nullptr;
     std::vector<int> filteredWorkouts = std::vector<int>{};
 
-    for(int i = 0; i < workout_options.size(); ++i) {
+    for(size_t i = 0; i < workout_options.size(); ++i) {
         if (workout_options[i].getType() == CARDIO) {
             if (cheapestCardio == nullptr) {
                 cheapestCardio = &(workout_options[i]);
